corrigindo at

+++ src/main/java/com/contatamento/contatos_api/Controllers/Dtos/ContatoResponse.java
@@ -3,5 +3,5 @@ package com.contatamento.contatos_api.Controllers.Dtos;
 public class ContatoResponse {
     public long id;
     public String nome;
-    public Long numero;
+    public Integer numero;
 }
