+++ src/main/java/com/contatamento/contatos_api/Controllers/Dtos/CreateContato.java
@@ -6,12 +6,11 @@ import com.contatamento.contatos_api.Models.Contato;
 public class CreateContato {
 
     public String nome;
-    public Long numero;
+    public Integer numero;
 
     public CreateContato() {}
 
     public Contato toContato(){
-
         return new Contato(this.nome, this.numero);
     }
 }
